+++ source/rejection.h
@@ -10,6 +10,8 @@ Int_t rejection(const Parameters &setting, std::vector< Event >& events );
 Int_t energyCut(const Parameters &setting, std::vector< Event >& events );
 
 Int_t PSDCut(const Parameters &setting, const UInt_t index, std::vector< Event >& events );
+
+Int_t simplePUR(const Parameters &setting, const UInt_t index, std::vector< Event >& events );
 // Int_t filterBad(const Parameters &setting, std::vector<Event> &events);
 
 #endif // REJECTION_H